requirements.txt
flask
ultralytics
pillow
torch
